Added check for redundant config change
Added a HEAD file that saves the latest configuration, in addition
to the configuration that is appended to the destination file.

When backup-running-config command is invoked, the fetched running
config is compared to the HEAD, aka last config. Only if they are
different is the destination file appended to and the HEAD is updated.